* 3.0:
  Fix PHP 7.1 related failures
  [VarDumper] Fix for 7.1
  fixed CS
  Added class existence check if is_subclass_of() fails in compiler passes
  Fix the DBAL session handler version check for Postgresql

+++ src/Symfony/Component/VarDumper/Cloner/VarCloner.php
@@ -315,7 +315,7 @@ class VarCloner extends AbstractCloner
             if (!empty($frame['line'])) {
                 ob_start();
                 debug_zval_dump($obj);
-                self::$hashMask = substr(ob_get_clean(), 17);
+                self::$hashMask = (int) substr(ob_get_clean(), 17);
             }
         }
